Moved files for the Contacts page.
